soc: ti: add helper functions for use across netcp drivers
The helper functions that format the descriptor and access the
common qmss queue related tasks are usable across all of netcp drivers.
So move these into knav_helpers under include/linux/soc/ti so that it can
be included in all of the knav netcp drivers outside of net/ethernet.
While at it, convert these into inline functions as they are used in
the packet processing functions.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>