net: ethernet: ti: prueth: [HACK] configure ptp master mac addr to ptp firmware
This is a HACK.

In order to support PTP time stamping using the PRU PRP firmware,
the firmware requires the Master clock (as result of the Best
Master Clock Algorithm) MAC address to be configured to a shared
memory location. The patch adds a hack that uses the set_dump
API in the struct ethtool_ops to let application to pass in the
MAC address of the Master clock.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>