prueth: fix debugfs directory to include pruss id
Current implementation of debugfs interface for hsr/prp breaks the
driver when the second instance of the driver is brought up on ICSS1.
This is because the path for hsr/prp debugfs directory doesn't take
into account the possibility of having a second instance. To allow
second instance and to keep backward compatibility with other
applications such as net-snmp that assumes the path, create the
ICSS2 instance with prueth-prp path and the ICSS1 instance with a
prueth-prp-0 path. This ensures existing application works fine.

Beware that currently the prueth driver works fine for eth4, but
not for eth5. As the SAN - PTP uses PRP in emac mode and BC requires
at least one Eth port functional, this should be fine.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>