hsr: remove the unnecessary kernel trace when hsr is offloaded
When hsr is offloaded, the offloaded h/w or f/w is expected to
consume the supervision frame. However there is no need for a
kernel trace here and frame should be silently discarded. Fix
this by removing the WARN_ONCE() and replace it with a comment
in the code.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Vitaly Andrianov <vitalya@ti.com>