net: hsr/prp: add vlan ctag filter support
This patch adds support for vlan ctag based filtering. The underlying
slave Ethenet device may be capable of filtering Ethernet packets based
on VLAN ID. This requires that when the VLAN interface is created over
an hsr/prp interface, it passes the vid information to the associated
slave Ethernet devices so that it updates the hardware filters to
filter Ethernet frames based on VID.

While adding this support, re-structure the code to re-use existing
function, hsr_prp_get_port(), to get slave port for mc and vlan filter
handling instead of iterate over the various ports to find slave
port in the handling code as is done currently.

Also use netdev_lockdep_set_classes() to allow nesting of lock as
is done in bonding/team driver.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>