net: driver: ti: prueth: handling HSR PTP event cut-through tx timestamp
When a HSR non-link-local PTP event is cut-through forward by the
HSR firmware, in addition to the rx timestamp, a tx timestamp is
also provided by the firmware. The tx timestamp is saved in the
shared redundant info area of the received skb before the skb
is passed to upper layer. Currently cut-through tx timestamp is
handled only for HSR PTP over L2 non-link-local event.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>