net: socket: forward extra timestamp of received skb to user space
Other than the normal receive timestamp, a packet received over
a redundant network port may have an additional timestamp associated
with. This is the transmit timestamp of the received packet over the
other redundant network port due to the cut-through forwarding of
underlying hardware device.

The additional timestamp, or the red_hwtstamps in the shared redundant
info area of the received skb, is forwarded to user space as
SCM_RED_TIMESTAMPING type control message.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>