hsr/prp: improvement and fixes for lan_id handling
Currently lan id is not set in supervisor frames. Also do a sanity
check for frames at the ingress for proper lan_id and update the
error stats counter when they don't match. This is needed to help
fix possible lan connection issues. For example, by observing the
stats counter, user can make sure the connection is done correctly. i.e
Slave-A should be connected to slave-A of the adjacent node for HSR ring
or PRP network A. Otherwise, cnt_rx_wrong_lan_a and cnt_rx_wrong_lan_b
will increment.

This also pass the rct ptr to prp_check_lsdu_size() to check LSDU size
instead of calling the skb_get_PRP_rct() again as the same is available
to the caller.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Vitaly Andrianov <vitalya@ti.com>