hsr/prp: fix protocol version number setting error
The protocol version number needs to be set before it is used. Otherwise
the behavior is undetermined. For example, in the case of PRP, the RCT
will have an incorrect LanID which that leads to receiving duplicate
frames, such as 3 duplicate ping responses (a total of 4 responses) for
each ping request generated.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Vitaly Andrianov <vitalya@ti.com>