net: ethernet: ti: cpts: add 1pps support using dmtimer
The CPTS hardware on AM57x doesn't have a pin out for PPS and to
work around that a DM timer based 1pps signal generation is done
using this patch. The DM timer is programmed to generate 100 msec
pulses and 9 out of 10 pulses are filtered out using pinmux control.
The driver implement a pps state machine and handle the events using
a kthread worker. This update also handles additional functions
such as

1) Register with boundary clock driver to allow bc functionality
   support
2) Enable/disable pps signal to a common bus to allow synchronization
   of slave clocks to the master clock (internal master, slave clock
   to remote master)
3) Generate its own 1 PPS signal output

Signed-off-by: Eric Ruei <e-ruei1@ti.com>
Signed-off-by: Vitaly Andrianov <vitalya@ti.com>