net: ethenet: prueth: introducing ptp time stamping support
This patch adds the support of IEEE-1588 ordinary clock (OC) on the
PRU-ICSS ethernet ports of TI's AM572x devices. The PTP clock is
implemented by the IEP driver and this patch adds linux ptp related
interfaces to kernel and adds hooks to tx and rx path for time stamping.
This initializes the iep driver and uses it to read the timestamp
from PRU.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>