net: ethernet: prueth: add support for rx interrupt pacing
Currently driver processes one interrupt per packet irrespective of
the size of frame. This causes excessive interrupt events for small
size frames at the host and results in lesser throughput performance.
This patch introduces rx interrupt pacing support using the ECAP
module in PRU ICSS. Firmware accumulates the frames in the ingress
queue and generates interrupt at configured timeout intervals using
ECAP. Additionally it also adds support for adaptive receive
which triggers an interrupt if the queue is empty and a new frame
arrives at the PRU. However this have performance impact and results
in less accumulation at the PRU. However if an application is
sensitive to latency, this may be used along with interrupt pacing.

Additonally this adds changes to make this compatible on other devices
that don't use ecap by not parsing the reg for ecap on those devices.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>