net: phy: dp83867: add workaround for incorrect default FLD threshold
When the Phy is bootstrapped to enable Fast Link Drop detect, the phy
link becomes unstable and goes up/down frequently. This is because the
FLD threashold value changes to 0x222 which is incorrect and should
be 0x221 as is the case when strapping is not used. As a workaround,
restore the value in bit 2:0 to 1.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>