net: hsr: rename functions to introduce prp support
There are many similarities with hsr and prp protocols. So a prp device
can be implemented by re-using functions from hsr code. So as a
preparatory patch, rename all common functions/defines/types with a
hsr_prp suffix. The hsr prefix is dropped from static function as they
are local to the file.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>