dt-bindings: net: ti,icssg-prueth: update for using icssg pair
Interposer card allows to pair two ICSSG to provide a pair of
PRU Ethernet ports. Each PRU handles a egress or ingress path
in this configuration. Introduce a new compatibility to
identify this particular hardware configuration so that driver
can be customized to use the two icssgs pair, one at egress and
another at ingress.  Introduce a new dt property, mii-gt-rt-paired
to points to the mii-g-rt of second icssg. Similarly introduce
another dt property, mii-rt-paired, to points to the mii-rt of
second iccsg.  This configuration assumes that first icssg is
ICSSG0 and second is ICSSG1.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>