net: hsr/prp: update to support lre rx and l2 forward offload
HSR LRE functionality as well L2 forwarding can be offloaded to hardware
(h/w) or firmware (f/w) if the device feature flag indicates the support.
When the LRE Ingress function is offloaded, the LRE function of
duplicating and tagging in egress direction is still handled in the
upper hsr/prp driver. At the ingress, the supervisor frames are processed
at h/w or f/w and is responsible for maintaining the node table for
remote nodes and statistics. So the node table related functions are
disabled in offloaded case. H/w or f/w is also expected to drop duplicate
of the received frames to host and forward only one copy to slave A or
Slave B depending on which interface frame came first. Non Host and
broadcast/multicast frames gets forwarded through L2 forwarder of the
HSR driver code if NETIF_F_HW_L2FW_DOFFLOAD is not supported.  If
supported, then no forwarding happens in the HSR L2 forwarder and the
same is assumed to be handled at h/w or f/w.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>