remoteproc/pruss: update pruss_get() to retrieve a PRUSS id
Update the pruss_get() function to take in an additional integer
pointer argument in which the PRUSS instance id is filled in and
provided back to the callers. This allows the drivers to add some
instance-specific logic/customization in their code, as the PRUSS
handle is not useful to build this logic.

The already existing usage within the PRU Ethernet driver has also
been updated accordingly, and this will cater to its need for
supporting switching between different Ethernet protocols  dynamically
per instance.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Aaron Kramer <a-kramer@ti.com>