HACK: net: prueth: use nt mac look up result only for rev 2.0 fw
The MAC address look up result in Node Table is not reliable on
rev 1.0 firmware.  A workaround was implemented in the driver to
discard the look up result from firmware and process the source mac
address for all frames for PRP and SV frames for HSR. This results
in unnecessary processing overhead for every packet for PRP and SV
frame for HSR in the rx packet processing code and in the NT worker.
Testing shows that this adds about 4% of CPU load in the worker code
and about 2% increase in CPU load in irq thread when 200 byte frames
are processed at 31Mbps which is significant. So add this workaround
conditionally for rev 1.0 firmware only.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Aaron Kramer <a-kramer@ti.com>