net: ethernet: ti: prueth: introduce iep ptp driver for prueth
This patch adds the support of IEEE-1588 ordinary clock (OC) on the
PRU-ICSS ethernet ports of TI's devices. The PRU-ICSS
includes an Industrial Ethernet Peripheral (IEP) unit. This unit
is a simple 64-bit timer (counter) whose functionality, in
conjunction with TI's PRU-ICSS industrial PRP firmware, is leveraged
in the support of IEEE-1588 PTP time stamping on the PRU-ICSS
ethernet ports.

This uses a different method to set the CMP1 register for SYNC
signal generation due to the 32-bit IEP wraparound. The CMP1 value must
be less than the IEP wraparound value (1 second), so it is truncated to
less than 1 second. Additionally, the SYNC signal must be reset *after*
IEP wraparound occurs or the SYNC pulse will trigger again because the
IEP value is still greater than CMP1. So SYNC is reset only during a
short interval immediately before the next SYNC pulse. This is
incomplete as this will succeed the majority of the time but fail
when CMP1 is a very low value because the SYNC reset may happen before
IEP wraparound in that case.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Aaron Kramer <a-kramer@ti.com>