remoteproc/pruss: restore iep as a memory node of pruss
The conversion of iep dt node to syscon breaks existing iep driver.
This patch restore the function to allow porting the iep driver as
an interim solution. Probably it may be a better option to use a
separate
dt node with compatibility for iep and convert the iep driver to a
platform driver. This will have to be resolved soon.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>
Signed-off-by: Aaron Kramer <a-kramer@ti.com>