net: driver: ti: prueth: add support of HSR PTP
This patch adds support of HSR PTP over L2 in prueth driver.
In particular, it adds support of hsr ptp packet processing
in prueth driver and enables forced 2-step in HSR firmware.

handling HSR PTP event cut-through tx timestamp:
When a HSR non-link-local PTP event is cut-through forward by the
HSR firmware, in addition to the rx timestamp, a tx timestamp is
also provided by the firmware. The tx timestamp is saved in the
shared redundant info area of the received skb before the skb
is passed to upper layer. Currently cut-through tx timestamp is
handled only for HSR PTP over L2 non-link-local event.

wait for cut-thru ptp timestamp:
In HSR, when there is a heavy load of network traffic, it may take
a little longer for the HSR firmware to return the cut-through tx
timestamp for ptp event messages such as SYNC. If the cut-thru tx
timestamp is not available on first read, then the driver
should save the ptp message locally and schedule a delayed work to
try to read the timestamp again later.

Signed-off-by: WingMan Kwok <w-kwok2@ti.com>
Signed-off-by: Aaron Kramer <a-kramer@ti.com>