net: ethernet: prueth: add PRP EMAC mode
This change adds support for PRP EMAC mode. This is configurable via
sysfs entry, where 'echo 1 >
/sys/devices/platform/pruss<ID>_eth/net/eth<if number>/prp_emac_mode'
enables, and writing 0 disables. This enables the 2 PRU-ICSS ports to
act independently as EMAC ports.
It is expected that this mode will be deprecated when the EMAC firmware
reaches feature parity with PRP firmware.

Signed-off-by: Aaron Kramer <a-kramer@ti.com>