net: ethernet: prueth: fix hsr/prp timestamping mem offset
The offset used for accessing timestamps should be the offset from the
start of the buffer descriptor memory, not from the start of each ports
buffer descriptor memory region, otherwise both ports are accessing
timestamps from the same array of timestamps. An identical correction
was needed in firmware to store the timestamps at the correct offsets
and not have each port overwrite the other.

Signed-off-by: Aaron Kramer <a-kramer@ti.com>