net: ethernet: prueth: fix storm prev enable for DualEMAC
The timer function for storm prevention checked that both ports are
configured before reseting the counter. In single EMAC case, only one
port may be configured so it should not check for both ports (this was
causing storm prevention to not function at all when only one port was
active). This change makes the check for both ports only applicable
to HSR/PRP case (the individual ports are still checked before
actual update).

Signed-off-by: Aaron Kramer <a-kramer@ti.com>