net: ethernet: icssg-prueth: add f/w command to set speed/duplex settings
Add command for changing speed or duplex settings at firmware.
Currently f/w detects the setting from RGMII cfg and same scheme has
some limitations in the firmware. To overcome that, Driver now call
emac_change_port_speed_duplex() to create the command and send it to
firmware for explicitly changing the speed/duplex setting at firmware.
The firmware requires that only one command to be sent to firmware
at a time. This requires that a cmd_in_progress is added to detect
the state of outstanding command. Also use spinlock to protect
against simultaneous access to send_command() function.

Following additional changes are done as well that are not directly
related to this.
 - Move the spinlock to protect write to h/w register close to to
   the actual function in emac_adjust_link().
 - Rename variable for shutdown_complete to cmd_complete as it is
   used for all commands.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>