net: ethernet: icssg-prueth: add 10M full duplex support
To support 10M Link speed at the ICSSG Ethernet ports, in band mode
status should be configured in RGMII CFG register. This allows PHY
to report speed/duplex settings in band to RGMII interface. However
not all PHY would support in-band status as it is an optional feature
described in the RGMII spec. So enable in band mode for 10M only if DT
property managed = "in-band-status" is set in the eth node. Otherwise
don't advertise 10M support.

ICSSG available on PG1 of the SoC doesn't have sufficient bits in the
MII TX IPG register for 10M. So firmware uses hard coded value of
IPG for 10M.

Signed-off-by: Murali Karicheri <m-karicheri2@ti.com>